Document, AnalyzedDocument 모듈 분리
