+++ keyext/__init__.py
@@ -1,17 +1,5 @@
-class Document(object):
-    def __init__(self, title, sentences):
-        super().__init__()
-        self.title = title
-        self.sentences = sentences
+__all__ = []
 
-
-class AnalyzedDocument(Document):
-    def __init__(self, document, vector, keywords, tags):
-        super().__init__(document.title, document.contents)
-
-        self.vector = vector
-        self.keywords = keywords
-        self.tags = tags
-
-from .preprocess import convert_to_document
-from .extraction import KeywordExtractor
+from ._model import Document, AnalyzedDocument
+from ._preprocess import convert_to_document
+from ._extraction import KeywordExtractor
